+++ packages/client/modules/pages/PageEntry.tsx
@@ -1,12 +1,12 @@
 import graphql from 'babel-plugin-relay/macro'
 import {type PreloadedQuery, usePreloadedQuery} from 'react-relay'
+import type {Page_viewer$key} from '../../__generated__/Page_viewer.graphql'
 import type {PageEntryQuery} from '../../__generated__/PageEntryQuery.graphql'
-import type {useTipTapPageEditor_viewer$key} from '../../__generated__/useTipTapPageEditor_viewer.graphql'
 import Page from './Page'
 import {PageNoAccess} from './PageNoAccess'
 
 interface Props {
-  viewerRef: useTipTapPageEditor_viewer$key | null
+  viewerRef: Page_viewer$key | null
   queryRef: PreloadedQuery<PageEntryQuery>
   isPublic?: boolean
   pageId: string